Moved interrupt code to an interface file
